# be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/payments": {
      "get": {
        "operationId": "PaymentsController_list",
        "parameters": [
          {
            "name": "reservationId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hotel_session": []
          }
        ],
        "summary": "List payments",
        "tags": [
          "payments"
        ]
      },
      "post": {
        "operationId": "PaymentsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hotel_session": []
          }
        ],
        "summary": "Record a manual payment",
        "tags": [
          "payments"
        ]
      }
    },
    "/api/v1/payments/{paymentId}/void": {
      "patch": {
        "operationId": "PaymentsController_void",
        "parameters": [
          {
            "name": "paymentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidPaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "hotel_session": []
          }
        ],
        "summary": "Void a manual payment",
        "tags": [
          "payments"
        ]
      }
    },
...
  },
  "info": {
    "title": "SUNSEA Hotel Management API",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "hotel_session"
      }
    },
    "schemas": {
...
      "ApiMetaDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "timestamp"
        ]
      },
...
      "PaymentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "reservationId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentType": {
            "type": "string",
            "enum": [
              "deposit",
              "settlement",
              "refund",
              "adjustment"
            ]
          },
          "amount": {
            "type": "number"
          },
          "method": {
            "type": "string",
            "enum": [
              "cash",
              "bank_transfer",
              "other"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "completed",
              "voided",
              "refunded"
            ]
          },
          "paidAt": {
            "type": "object",
            "format": "date-time",
            "nullable": true
          },
          "note": {
            "type": "object",
            "nullable": true
          },
          "voidReason": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "reservationId",
          "paymentType",
          "amount",
          "method",
          "status"
        ]
      },
      "PaymentListResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentResponseDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMetaDto"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "CreatePaymentDto": {
        "type": "object",
        "properties": {
          "reservationId": {
            "type": "string",
            "format": "uuid"
          },
          "paymentType": {
            "type": "string",
            "enum": [
              "deposit",
              "settlement",
              "refund",
              "adjustment"
            ]
          },
          "amount": {
            "type": "number",
            "example": 500000
          },
          "method": {
            "type": "string",
            "enum": [
              "cash",
              "bank_transfer",
              "other"
            ]
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "reservationId",
          "paymentType",
          "amount",
          "method"
        ]
      },
      "PaymentEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/PaymentResponseDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMetaDto"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "VoidPaymentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "reason"
        ]
      }
    }
  }
}